docs: create rpmsg design file in readthedocs to use breathe references
the rpmsg-design file is a markdown (md) file in open-amp repository
which uses code within the file to demonstrate APIs.
Review comments were to use doxygen linking, so use breathe and doxylink
in a new file within readthedocs to achieve the same references but
through embedding which will maintain itself as long as doxygen comments
are maintained in open-amp repo.